feat: update SvgRendererV1 and TokenURIRendererV1

+++ src/interfaces/IGame.sol
@@ -8,7 +8,7 @@ interface IGame {
 
     function cardDurability(uint256 tokenId) external view returns (uint8);
 
-    function cardPower(uint256 tokenId) external view returns (uint8);
+    function cardPower(uint256 tokenId) external view returns (uint32);
 
     function cardLevel(uint256 tokenId) external view returns (uint8);
